Rename and remove logs
